+++ packages/components/src/Provider/index.ts
@@ -1,4 +1,5 @@
 export { useTheme, ThemeProvider } from '@marigold/system';
+export { I18nProvider } from '@react-aria/i18n';
 
 export * from './MarigoldProvider';
 export * from './OverlayContainerProvider';