+++ themes/theme-b2b/src/components/DateField.styles.ts
@@ -1,6 +1,11 @@
 import { ThemeComponent, cva } from '@marigold/system';
 
-import { inputBackground, inputBox, inputSpacing } from './Input.styles';
+import {
+  inputBackground,
+  inputBox,
+  inputPlaceholder,
+  inputSpacing,
+} from './Input.styles';
 
 export const DateField: ThemeComponent<'DateField'> = {
   field: cva([
@@ -9,11 +14,11 @@ export const DateField: ThemeComponent<'DateField'> = {
     inputBackground,
     inputSpacing,
     'h-[34px]',
-    'data-[hovered]:border-border-hover',
+    'data-[hovered]:border-border-base-hover',
     // inputDisabled('data-[disabled]') color tokens are not working with in this component
-    'data-[disabled]:bg-bg-disabled data-[disabled]:border-border-disabled data-[disabled]:text-text-disabled data-[disabled]:cursor-not-allowed',
+    'data-[disabled]:bg-bg-base-disabled data-[disabled]:border-border-base-disabled data-[disabled]:text-text-base-disabled data-[disabled]:cursor-not-allowed',
     // inputFocus('data-[focus-within]') is not working with this component
-    'data-[focus-within]:border-border-focus data-[focus-within]:outline-outline-focus',
+    'data-[focus-within]:outline-outline-focus',
     'group-data-[invalid]/field:border-border-error',
   ]),
   segment: cva([
@@ -21,9 +26,10 @@ export const DateField: ThemeComponent<'DateField'> = {
     'text-center leading-none outline-0',
     '[data-type="literal]"' && 'p-[0.75px]',
     'rounded-sm',
-    'focus:bg-bg-focus focus:text-white',
-    'aria-[readonly]:focus:border-border-light aria-[readonly]:focus:outline-none',
-    'aria-[readonly]:bg-bg-transparent aria-[readonly]:text-text-disabled',
+    'focus:bg-bg-accent focus:text-text-inverted',
+    'aria-[readonly]:focus:border-border-base aria-[readonly]:focus:outline-none',
+    'aria-[readonly]:text-text-base-disabled aria-[readonly]:bg-transparent',
+    inputPlaceholder(),
   ]),
   action: cva(''),
 };