+++ themes/theme-b2b/src/components/Input.styles.ts
@@ -1,21 +1,25 @@
 import { ThemeComponent, cva } from '@marigold/system';
 
 export const inputBox =
-  'border-border-light rounded-sm border outline-none outline-offset-[-1px]';
+  'border-border-base rounded-sm border outline-none outline-offset-[-1px]';
 export const inputBackground = 'bg-bg-surface';
 export const inputSpacing = 'px-2';
 
 export const inputFocus = (selector: string = 'focus') =>
-  `${selector}:border-border-focus ${selector}:outline-outline-focus`;
+  `${selector}:outline-outline-focus`;
 
 export const inputDisabled = (selector: string = 'disabled') =>
-  `${selector}:bg-bg-disabled ${selector}:text-text-disabled ${selector}:hover:border-border-disabled ${selector}:border-border-disabled ${selector}:cursor-not-allowed`;
+  `${selector}:bg-bg-base-disabled ${selector}:text-text-base-disabled ${selector}:hover:border-border-base-disabled ${selector}:border-border-base-disabled ${selector}:cursor-not-allowed`;
 
 export const inputError = (selector: string = 'error') =>
-  `${selector}:border-border-error ${selector}:outline-outline-error ${selector}:-outline-offset-1`;
+  `${selector}:border-border-error`;
 
 export const inputHover = (selector: string = 'hover') =>
-  `${selector}:border-border-hover`;
+  `${selector}:border-border-base-hover`;
+
+export const inputPlaceholder = (
+  selector: string = 'data-[placeholder=true]'
+) => `${selector}:text-text-base-disabled`;
 
 export const Input: ThemeComponent<'Input'> = {
   icon: cva('left-2'),
@@ -28,8 +32,9 @@ export const Input: ThemeComponent<'Input'> = {
     inputHover(),
     inputError('group-error/field'),
     inputSpacing,
+    inputPlaceholder(),
     'leading-[2]',
-    'read-only:focus:border-border-light read-only:focus:outline-none',
+    'read-only:focus:border-border-base read-only:focus:outline-none',
     // Extra padding for when an icon/action is present
     'group-data-[icon]/input:pl-8',
   ]),