+++ themes/theme-b2b/src/components/Label.styles.ts
@@ -1,8 +1,8 @@
 import { ThemeComponent, cva } from '@marigold/system';
 
 export const Label: ThemeComponent<'Label'> = {
-  container: cva('group-disabled/field:text-text-disabled text-sm '),
+  container: cva('group-disabled/field:text-text-base-disabled text-sm '),
   indicator: cva(
-    'group-required/field:text-text-required group-required/field:block'
+    'group-required/field:text-text-error group-required/field:block'
   ),
 };