+++ themes/theme-b2b/src/components/Select.styles.ts
@@ -7,6 +7,7 @@ import {
   inputError,
   inputFocus,
   inputHover,
+  inputPlaceholder,
   inputSpacing,
 } from './Input.styles';
 
@@ -20,8 +21,9 @@ export const Select: ThemeComponent<'Select'> = {
     inputHover(),
     inputBackground,
     inputSpacing,
+    inputPlaceholder('[&>[data-placeholder=true]]'),
     'appearance-none leading-8',
     'cursor-pointer',
-    'aria-expanded:bg-bg-neutral ',
+    'aria-expanded:bg-bg-base-active ',
   ]),
 };