+++ themes/theme-core/src/preset.ts
@@ -1,5 +1,4 @@
-import { createPreset } from '@marigold/theme-preset';
-import { flattenObject } from '@marigold/theme-preset/src/utils';
+import { createPreset, flattenObject } from '@marigold/theme-preset';
 
 import { screens } from './screens';
 import { colors, height, shadow } from './tokens';